Move the code to their corresponding go file
